pyspark-translation-example/pyproject.toml
[project]
name = "neo4j-sql-translation"
version = "0.2.0"
description = "Test Neo4j JDBC SQL-to-Cypher translation locally with PySpark"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyspark>=3.5.0,<4.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
# Main entry point - runs all SQL translation and schema discovery tests
test-sql = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include both the main module and the test package
packages = ["main.py", "neo4j_sql_tests"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
